backend/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gitdb==4.0.11
GitPython==3.1.41
idna==3.7
psycopg2-binary==2.9.9
PyJWT==2.8.0
pymemcache==4.0.0
python-dotenv==1.0.1
requests==2.31.0
smmap==5.0.1
sqlparse==0.4.4
urllib3==2.2.1
